!171 check usage when deleting attribute value * 删除AttributeValue前检查是否有Product使用

+++ beike/Admin/Repositories/AttributeRepo.php
@@ -99,6 +99,10 @@ class AttributeRepo
 
     public static function deleteValue($id)
     {
+        $productIds = ProductAttribute::query()->where('attribute_value_id', $id)->pluck('product_id')->toArray();
+        if ($productIds) {
+            throw new \Exception(trans('admin/attribute.error_cannot_delete_product_used', ['product_ids' => implode(', ', $productIds)]));
+        }
         AttributeValue::query()->findOrFail($id)->delete();
     }
